pyproject.toml
[project]
name = "marketing-intelligence-agent"
version = "0.1.0"
description = "AI-powered marketing intelligence agent with LangGraph, RAG, and Streamlit UI."
authors = [{ name = "Simon" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-community>=0.3.0",
    "qdrant-client>=1.9.0",
    "sentence-transformers>=2.2.0",
    "pandas>=2.0.0",
    "polars>=0.20.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "streamlit>=1.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "groq>=0.4.0",
    "tenacity>=8.0.0",
    "langfuse>=2.0.0",
    "ragas>=0.1.0",
    "plotly>=5.18.0",
    "mangum>=0.17.0",
    # AWS integrations
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    # PDF export and enhanced UI
    "fpdf2>=2.7.0",
    "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-requests>=2.32.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

